workbench: break perspective selector into its own widget
Much like the old sidebar selector, this works as a GtkMenuButton with a
popover of the perspectives (and is mostly self-contained).

The workbench can add the widget to the header bar in a simiarl way to how
we would expect plugins to add to the header bar (although they need to
call API to do it, and the workbench can do it from .ui files).