workbench: add sizing workarounds for centering omnibar
This is less than ideal, because it's the result of 2 bugs in gtk+ size
requests that we are working around.

The first, is the lack of hexpand support in left/right of the headerbar.
The second, is that the center widget + sizing of neighbors is not
respected when a GtkBox is packed as the custom title.

Where this will be apparent, is when the content is compressed and there
are uneven buttons on the left/right. You'll see spacing enforced for the
size group contained widgetry, which will be rather annoying im presume.

Either way, this works for now and allows us to move forward.