flatpak: Discover flatpak manifests and add them as runtimes
When a repo includes a flatpak manifest, GNOME Builder should know
how to use it with flatpak-builder to build the project's dependencies
so the build will be more likely to work. This commit teaches
GbpFlatpakRuntimeProvider how to identify manifests (by file name and
doing a sanity check on the contents) and add them to the list of
runtimes.

https://bugzilla.gnome.org/show_bug.cgi?id=773764