runner: Discover build targets after the build completes
When the user creates a new project and hasn't done a build yet the run
button fails because it tries to discover the build targets (by running
make) before doing a build but the ~/.cache/gnome-builder/builds/...
folder hasn't been created yet. This commit moves that operation to
after the build so that it succeeds on first run.

https://bugzilla.gnome.org/show_bug.cgi?id=773036